requirements/dev.txt
apns2==0.5.0
Babel==2.7.0
CacheControl==0.12.6
cachetools==3.1.1
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.1.12
django-cors-headers==3.1.1
django-countries==5.4
django-dbbackup==3.2.0
django-environ==0.4.5
django-extensions
django-money==1.0
django-phonenumber-field==3.0.1
django-rest-swagger==2.1.2
django-storages==1.8
django-twilio==0.10.0
djangorestframework==3.10.2
drf-access-policy==0.5.1
drf-yasg==1.17.0
environ==1.0
firebase==3.0.1
firebase-admin==3.2.1
gcloud==0.18.3
GeoFire-Python==0.3
geographiclib==1.50
geoindex==0.0.1
geopy==1.21.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.8.2
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.17.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jwcrypto==0.6.0
jws==0.1.3
MarkupSafe==1.1.1
msgpack==0.6.2
oauth2client==4.1.2
openapi-codec==1.3.2
packaging==19.2
phonenumbers==8.10.17
Pillow==6.2.1
protobuf==3.11.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
py-moneyed==0.8.0
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.6
pyfcm==1.4.7
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.2
Pyrebase4==4.3.0
PySocks==1.7.0
python-geohash==0.8.5
python-jwt==2.0.1
pytz==2019.2
pywebpush==1.10.1
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
sseclient==0.0.24
twilio==6.29.4
uritemplate==3.0.0
urllib3==1.25.3
